refactor Elastic Index JSON parsing to its own class
