+++ modules/services/ogc-api-records/src/main/java/org/fao/geonet/ogcapi/records/util/CollectionInfoBuilder.java
@@ -10,9 +10,7 @@ import static java.util.Arrays.asList;
 import static org.fao.geonet.ogcapi.records.util.LinksItemsBuilder.getHref;
 
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import javax.servlet.http.HttpServletRequest;
 import lombok.extern.slf4j.Slf4j;
@@ -22,19 +20,14 @@ import org.fao.geonet.domain.Source;
 import org.fao.geonet.domain.SourceType;
 import org.fao.geonet.ogcapi.records.controller.model.CollectionInfo;
 import org.fao.geonet.ogcapi.records.controller.model.CrsEnum;
-import org.fao.geonet.ogcapi.records.model.OgcApiContact;
 import org.fao.geonet.ogcapi.records.model.OgcApiExtent;
-import org.fao.geonet.ogcapi.records.model.OgcApiLanguage;
 import org.fao.geonet.ogcapi.records.model.OgcApiLink;
 import org.fao.geonet.ogcapi.records.model.OgcApiSpatialExtent;
-import org.fao.geonet.ogcapi.records.model.OgcApiTemporalExtent;
-import org.fao.geonet.ogcapi.records.model.OgcApiTheme;
 import org.fao.geonet.ogcapi.records.service.RecordService;
 import org.fao.geonet.repository.SettingRepository;
 import org.fao.geonet.repository.SourceRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
 
 @Service
 @Slf4j(topic = "org.fao.geonet.ogcapi.records")
@@ -49,46 +42,14 @@ public class CollectionInfoBuilder {
   @Autowired
   private SettingRepository settingRepository;
 
+  @Autowired
+  ElasticIndexJson2CollectionInfo elasticIndexJson2CollectionInfo;
+
   public CollectionInfoBuilder() {
 
   }
 
-  /**
-   * todo - be language aware (send in desired language).  Move to utility class.
-   *
-   * @param jsonNode json node for the potentially multi-language string
-   * @return "correct" language value.
-   */
-  public static String getLangString(Object jsonNode) {
-    if (jsonNode == null) {
-      return null;
-    }
-    if ((jsonNode instanceof String)) {
-      return (String) jsonNode;
-    }
-    if (jsonNode instanceof Map) {
-      var map = (Map<String, Object>) jsonNode;
-      if (map.containsKey("default")) {
-        return map.get("default").toString();
-      }
-      return null;
-    }
-    return null;
-  }
 
-  /**
-   * Simple utility class to get a JSON value as a string.
-   *
-   * @param o json object
-   * @return null or o.toString()
-   */
-  public static String getAsString(Object o) {
-    if (o == null) {
-      return null;
-    }
-    var result = o.toString();
-    return result;
-  }
 
 
   /**
@@ -149,131 +110,14 @@ public class CollectionInfoBuilder {
         format, collectionUri.toString(), language, configuration);
     linkList.forEach(collectionInfo::addLinksItem);
 
-    var linkedServiceRecord = recordService.getLinkedServiceRecord(request, source);
-    injectLinkedServiceRecordInfo(collectionInfo, linkedServiceRecord);
+    var linkedServiceRecord =
+        recordService.getLinkedServiceRecord(request, source);
+    elasticIndexJson2CollectionInfo.injectLinkedServiceRecordInfo(collectionInfo,
+        linkedServiceRecord);
 
     return collectionInfo;
   }
 
-  /**
-   * inject the "extra" info from the LinkedServiceRecord into the CollectionInfo.
-   *
-   * @param collectionInfo      collection metadata we've gathered so far (usually not much)
-   * @param linkedServiceRecord JSON of the linked Service record (GN's DB "source"
-   *                            "serviceRecord")
-   */
-  private void injectLinkedServiceRecordInfo(CollectionInfo collectionInfo,
-      Map<String, Object> linkedServiceRecord) {
-    if (linkedServiceRecord == null) {
-      return;
-    }
 
-    //override title from attached ServiceRecord
-    var title = getLangString(linkedServiceRecord.get("resourceTitleObject"));
-    if (title != null) {
-      collectionInfo.setTitle(title);
-    }
-
-    //override description (abstract) from attached ServiceRecord
-    var desc = getLangString(linkedServiceRecord.get("resourceAbstractObject"));
-    if (desc != null) {
-      collectionInfo.setDescription(desc);
-    }
-
-    var contacts = linkedServiceRecord.get("contact");
-    if (contacts != null && linkedServiceRecord.get("contact") instanceof List) {
-      var cs = (List) contacts;
-      collectionInfo.setContacts(new ArrayList<>());
-      for (var contactMap : cs) {
-        var contact = OgcApiContact.fromIndexMap((Map<String, Object>) contactMap);
-        collectionInfo.getContacts().add(contact);
-      }
-    }
-
-    OgcApiSpatialExtent spatialExtent = null;
-    var mySpatialExtent = linkedServiceRecord.get("geom");
-    if (mySpatialExtent != null
-        && mySpatialExtent instanceof List) {
-      mySpatialExtent = ((List) mySpatialExtent).get(0);
-      spatialExtent = OgcApiSpatialExtent.fromGnIndexRecord((Map<String, Object>) mySpatialExtent);
-    }
-
-    OgcApiTemporalExtent temporalExtent = null;
-    var myTemporalExtent = linkedServiceRecord.get("resourceTemporalDateRange");
-    if (myTemporalExtent != null && myTemporalExtent instanceof List) {
-      myTemporalExtent = ((List) myTemporalExtent).get(0);
-      temporalExtent = OgcApiTemporalExtent.fromGnIndexRecord(
-          (Map<String, Object>) myTemporalExtent);
-    }
-
-    OgcApiExtent extent = new OgcApiExtent(spatialExtent, temporalExtent);
-    if (spatialExtent != null || temporalExtent != null) {
-      collectionInfo.setExtent(extent);
-    }
-
-    var myCrss = linkedServiceRecord.get("coordinateSystem");
-    if (myCrss != null && myCrss instanceof List) {
-      var crss = (List) myCrss;
-      collectionInfo.setCrs(new ArrayList<>());
-      for (var crs : crss) {
-        collectionInfo.getCrs().add(crs.toString());
-      }
-    }
-
-    var createDate = linkedServiceRecord.get("createDate");
-    if (createDate != null && StringUtils.hasText(createDate.toString())) {
-      collectionInfo.setCreated(createDate.toString());
-    }
-
-    var updateDate = linkedServiceRecord.get("changeDate");
-    if (updateDate != null && StringUtils.hasText(updateDate.toString())) {
-      collectionInfo.setUpdated(updateDate.toString());
-    }
-
-    var myTags = linkedServiceRecord.get("tag");
-    if (myTags != null && myTags instanceof List) {
-      var tags = (List) myTags;
-      collectionInfo.setKeywords(new ArrayList<>());
-      for (var tag : tags) {
-        collectionInfo.getKeywords().add(getLangString(tag));
-      }
-    }
-
-    //index record doesn't contain "otherlanguages"
-    var myMainLang = linkedServiceRecord.get("mainLanguage");
-    if (myMainLang != null && StringUtils.hasText(myMainLang.toString())) {
-      collectionInfo.setLanguage(new OgcApiLanguage(myMainLang.toString()));
-    }
-
-    var myOtherLangs = linkedServiceRecord.get("otherLanguage");
-    if (myOtherLangs != null && myOtherLangs instanceof List) {
-      var langs = (List) myOtherLangs;
-      collectionInfo.setLanguages(new ArrayList<>());
-      for (var lang : langs) {
-        collectionInfo.getLanguages().add(new OgcApiLanguage(lang.toString()));
-      }
-    }
-
-    var myThemeKeywords = linkedServiceRecord.get("allKeywords");
-    collectionInfo.setThemes(OgcApiTheme.parseElasticIndex((Map<String, Object>) myThemeKeywords));
-
-    //its a bit unclear what to do here - there's a big difference between MD record,
-    // Elastic Index JSON versus what's expected in the ogcapi license field.  We do the simple
-    // action
-    var myLicense = linkedServiceRecord.get("MD_LegalConstraintsUseLimitationObject");
-    if (myLicense != null && myLicense instanceof List) {
-      var myLiceseList = (List) myLicense;
-      if (myLiceseList.size() > 0) {
-        var license = getLangString(myLiceseList.get(0));
-        collectionInfo.setLicense(license);
-      }
-      if (myLiceseList.size() > 1) {
-        var license2 = getLangString(myLiceseList.get(1));
-        collectionInfo.setRights(license2);
-      }
-    }
-
-
-  }
 
 }
